fix(selectors,actions): move back validateLink selector (from Editor to Project) and rename arr2obj to indexById
